requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-emit-project --frozen --output-file=requirements.txt --no-dev --all-extras
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.5
    # via fastmcp
beartype==0.22.7
    # via
    #   py-key-value-aio
    #   py-key-value-shared
cachetools==6.2.2
    # via py-key-value-aio
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
cfgv==3.5.0
    # via pre-commit
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
coverage==7.12.0
    # via pytest-cov
cryptography==46.0.3
    # via
    #   authlib
    #   pyjwt
cyclopts==4.3.0
    # via fastmcp
diskcache==5.6.3
    # via py-key-value-aio
distlib==0.4.0
    # via virtualenv
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22.3
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.1
    # via fastmcp
fastmcp==2.13.2
    # via ibm-watsonxdata-mcp-server
filelock==3.20.1
    # via
    #   ibm-watsonxdata-mcp-server
    #   virtualenv
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   ibm-watsonxdata-mcp-server
    #   mcp
    #   respx
httpx-sse==0.4.3
    # via mcp
ibm-cloud-sdk-core==3.24.2
    # via ibm-watsonxdata-mcp-server
identify==2.6.15
    # via pre-commit
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
iniconfig==2.3.0
    # via pytest
jsonschema==4.25.1
    # via mcp
jsonschema-path==0.3.4
    # via fastmcp
jsonschema-specifications==2025.9.1
    # via jsonschema
librt==0.6.3
    # via mypy
markdown-it-py==4.0.0
    # via rich
mcp==1.23.1
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
mypy==1.19.0
    # via ibm-watsonxdata-mcp-server
mypy-extensions==1.1.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
openapi-pydantic==0.5.1
    # via fastmcp
opentelemetry-api==1.38.0
    # via
    #   ibm-watsonxdata-mcp-server
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-instrumentation==0.59b0
    # via opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-httpx==0.59b0
    # via ibm-watsonxdata-mcp-server
opentelemetry-sdk==1.38.0
    # via ibm-watsonxdata-mcp-server
opentelemetry-semantic-conventions==0.59b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
opentelemetry-util-http==0.59b0
    # via opentelemetry-instrumentation-httpx
packaging==25.0
    # via
    #   opentelemetry-instrumentation
    #   pytest
pathable==0.4.4
    # via jsonschema-path
pathspec==0.12.1
    # via mypy
pathvalidate==3.3.1
    # via py-key-value-aio
platformdirs==4.5.0
    # via
    #   fastmcp
    #   virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.5.0
    # via ibm-watsonxdata-mcp-server
py-key-value-aio==0.3.0
    # via fastmcp
py-key-value-shared==0.3.0
    # via py-key-value-aio
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   fastmcp
    #   ibm-watsonxdata-mcp-server
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via
    #   ibm-watsonxdata-mcp-server
    #   mcp
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via
    #   ibm-cloud-sdk-core
    #   mcp
pyperclip==1.11.0
    # via fastmcp
pytest==9.0.1
    # via
    #   ibm-watsonxdata-mcp-server
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-env
pytest-asyncio==1.3.0
    # via ibm-watsonxdata-mcp-server
pytest-cov==7.0.0
    # via ibm-watsonxdata-mcp-server
pytest-env==1.2.0
    # via ibm-watsonxdata-mcp-server
python-dateutil==2.9.0.post0
    # via ibm-cloud-sdk-core
python-dotenv==1.2.1
    # via
    #   fastmcp
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pywin32==311 ; sys_platform == 'win32'
    # via mcp
pyyaml==6.0.3
    # via
    #   jsonschema-path
    #   pre-commit
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   ibm-cloud-sdk-core
    #   jsonschema-path
respx==0.22.0
    # via ibm-watsonxdata-mcp-server
rich==14.2.0
    # via
    #   cyclopts
    #   fastmcp
    #   rich-rst
rich-rst==1.3.2
    # via cyclopts
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
ruff==0.14.7
    # via ibm-watsonxdata-mcp-server
six==1.17.0
    # via python-dateutil
sse-starlette==3.0.3
    # via mcp
starlette==0.50.0
    # via mcp
structlog==25.5.0
    # via ibm-watsonxdata-mcp-server
tomli==2.3.0 ; python_full_version <= '3.11'
    # via coverage
types-requests==2.32.4.20250913
    # via ibm-watsonxdata-mcp-server
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   mcp
    #   mypy
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   py-key-value-shared
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
urllib3==2.6.2
    # via
    #   ibm-cloud-sdk-core
    #   ibm-watsonxdata-mcp-server
    #   requests
    #   types-requests
uvicorn==0.38.0
    # via
    #   fastmcp
    #   mcp
virtualenv==20.35.4
    # via pre-commit
websockets==15.0.1
    # via fastmcp
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
zipp==3.23.0
    # via importlib-metadata
